Fixed migration

+++ src/database/migrations/2016_10_21_095713_blogcomments.php
@@ -13,23 +13,25 @@ class Blogcomments extends Migration
      */
     public function up()
     {
-    	Schema::create('blogcomments', function (Blueprint $table) {
-    		$table->increments('id');
-    		$table->integer('blog_id')->unsigned()->default(0);
-    		$table->integer('user_id')->unsigned()->default(0);
-    		$table->text('body');
-    		$table->timestamps();
-    	
-    		$table->foreign('blog_id')
-    			->references('id')
-    			->on('blogs')
-    			->onDelete('cascade');
-    		
-    		$table->foreign('user_id')
-    			->references('id')
-    			->on('users')
-    			->onDelete('cascade');
-    	});
+        if (!Schema::hasTable('blogcomments')) {
+            Schema::create('blogcomments', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('blog_id')->unsigned()->default(0);
+                $table->integer('user_id')->unsigned()->default(0);
+                $table->text('body');
+                $table->timestamps();
+
+                $table->foreign('blog_id')
+                    ->references('id')
+                    ->on('blogs')
+                    ->onDelete('cascade');
+
+                $table->foreign('user_id')
+                    ->references('id')
+                    ->on('users')
+                    ->onDelete('cascade');
+            });
+        }
     }
 
     /**
